abstract away chart details to TimeChart
