+++ devfiles/src/classes/widgets/LineChart.tsx
@@ -3,42 +3,12 @@ import Sidebar from '../Sidebar.js';
 import ExportFileType from './ExportFileType.js';
 import Plot from 'react-plotly.js';
 import { DayGrouping, YGrouping } from './Grouping.js';
+import TimeChart from './TimeChart.js';
 
-export default class LineChart extends Widget {
-    public generateSidebar(): Sidebar {
-        return new Sidebar([]);
-    }
-    public render(): JSX.Element {
-        const grouping = new DayGrouping(this.panel.dataFilterer, YGrouping.SpeciesGroup);
-        const plotLayout = {
-            width: 400,
-            height: 210,
-            margin: {
-                l: 30,
-                r: 30,
-                b: 50,
-                t: 65
-            }
+export default class LineChart extends TimeChart {
+    public chartSpecificLayout(): object {
+        return {
+            type: 'scatter'
         };
-        const { traces, layout } = grouping.getChart(
-            {
-                type: 'scatter'
-            },
-            plotLayout
-        );
-        const plotConfig = {
-            //staticPlot: true,
-            modeBarButtonsToRemove: ['zoomIn2d', 'zoomOut2d']
-        };
-        return <Plot data={traces} layout={layout} config={plotConfig} />;
-    }
-    public delete(): void {
-        //throw new Error('Method not implemented.');
-    }
-    public clone(): Widget {
-        throw new Error('Method not implemented.');
-    }
-    public export(fileType: ExportFileType): void {
-        throw new Error('Method not implemented.');
     }
 }
