+++ devfiles/src/classes/widgets/TimeChart.tsx
@@ -1,0 +1,54 @@
+import Panel from '../Panel';
+import Plot from 'react-plotly.js';
+import Sidebar from '../Sidebar';
+import MutuallyExclusiveSelector from '../options/MutuallyExclusiveSelector';
+import ExportFileType from './ExportFileType';
+import { ContinuousMonthGrouping, DayGrouping, YGrouping } from './Grouping';
+import Widget from './Widget';
+import WidgetConfig from './WidgetConfig';
+
+export default abstract class TimeChart extends Widget {
+    public constructor(panel: Panel, config: WidgetConfig) {
+        super(panel, config);
+        this.generateOptions();
+    }
+    public generateOptions(): void {
+        const yAxisSelector = new MutuallyExclusiveSelector(
+            this.panel,
+            'Y Axis',
+            ['Vulnerability Status', 'Species Group', 'Species'],
+            'Species'
+        );
+        this.options = [yAxisSelector];
+    }
+    public generateSidebar(): Sidebar {
+        return new Sidebar(this.options);
+    }
+    public abstract chartSpecificLayout(): object;
+    public render(): JSX.Element {
+        const grouping = new DayGrouping(this.panel.dataFilterer, YGrouping.Species);
+        const plotLayout = {
+            width: 400,
+            height: 210,
+            margin: {
+                l: 30,
+                r: 30,
+                b: 50,
+                t: 65
+            }
+        };
+        const { traces, layout } = grouping.getChart(this.chartSpecificLayout(), plotLayout);
+        const plotConfig = {
+            //staticPlot: true,
+            modeBarButtonsToRemove: ['zoomIn2d', 'zoomOut2d']
+        };
+        return <Plot data={traces} layout={layout} config={plotConfig} />;
+    }
+    public delete(): void {}
+    public clone(): Widget {
+        throw new Error('Method not implemented.');
+    }
+    public export(fileType: ExportFileType): void {
+        throw new Error('Method not implemented.');
+    }
+}
